requirements.txt
numpy~=1.22.1
pandas~=1.4.0
matplotlib~=3.5.1
seaborn~=0.11.2
scikit-learn~=1.0.2
Flask~=2.0.2
requests~=2.27.1
Pillow~=9.0.0
keras~=2.8.0
tensorflow~=2.8.0rc1
google-api-python-client~=2.37.0
click~=8.0.3
python-dotenv~=0.19.2
pytest~=7.0.1